아이템 시스템 수정중

+++ Monster.h
@@ -28,8 +28,8 @@ public:
 
 	//int maxEnergy;
 	int power;
-	int nowEnergy;
-	int maxEnergy = 0;
+	float nowEnergy;
+	float maxEnergy = 0;
 	int monsterNum;
 	bool isMove = false;
 	//bool isAlive = true;
